Separate the grouping and labelling button on a separate pane as well
